WebScrapping/requirement.txt
selenium==4.6.0
behave
allure_behave
